+++ changelog.md
@@ -2,5 +2,10 @@
 
 ![TARDIS Refined](https://wiki.tardisrefined.net/TARDIS-Refined-Wiki/tardis_refined_v2_1.png)
 
+#### Interiors
+- Added Interior: Hartnell 
+- Added Interior: Pertwee 
+
 #### Enhancements
 - Enhancement: Added tooltips to various screens to indicate which mod added content
+- Enhancement: Players can now sleep in the TARDIS (No spawnpoint setting)